requirements.txt
aiohttp
albumentations==0.4.3
async_timeout
asyncio
autopep8
boto3
bs4==0.0.1
cryptography
cython
datasets==2.8.0
devpi-server==4.8.1
diffusers
einops==0.3.0
facenet_pytorch
fairscale==0.4.6
flask-cors==3.0.10
flask==2.0.1
fuzzywuzzy
glfw
gradio
gym[mujoco]
gym[robotics]
h5py
imageio
imageio-ffmpeg==0.4.4
imageio==2.9.0
invisible-watermark
ipdb
ipykernel
ipywidgets
jsonlines
jupyter
k-diffusion==0.0.16
keras
kornia==0.6
lmdb
markdown-it-py>=1.0
matplotlib
moviepy
# mpi4py==3.0.0
mujoco-py==1.50.1.56
notebook
omegaconf==2.3.0
open_clip_torch==2.20.0
openai
opencv-python==4.5.5.64
paddleocr>=2.0.1
Pillow==9.5.0
pudb==2019.2
protobuf>=3.20.0
pydevd-pycharm~=212.5457.59
pylint
pyspng==0.1.0
python-Levenshtein
pytorch-fid==0.2.1
pytorch-lightning==2.0.7
pytorch-pretrained-BERT
pyyaml
requests
scikit-image
scikit-learn
scipy
seaborn==0.11.2
selenium==4.3.0
six
slackclient
stanford-corenlp
streamlit>=0.73.1
taming-transformers==0.0.1
tensorboardX
test-tube>=0.7.5
timm==0.5.4
torch-fidelity==0.3.0
torch>=1.10.1
torchmetrics
tqdm
transformers==4.20.1
virtualenv
wandb
